Updated Sampling rate with GET FFT

+++ 1. IDE_PLUS Firmwares/butterfly_firmware/RawDataState.cpp
@@ -20,7 +20,7 @@ namespace FeatureStates {
     double isr_startTime;
     double isr_stopTime ;
     double elapsedTime;
-    double outputFrequency;
+    int outputFrequency;
     int isrCount =0;
