+++ 1. IDE_PLUS Firmwares/butterfly_firmware/RawDataState.h
@@ -29,7 +29,7 @@ namespace FeatureStates {
     extern double isr_startTime;
     extern double isr_stopTime;
     extern double elapsedTime;
-    extern double outputFrequency;
+    extern int outputFrequency;
     extern int isrCount; 
 }
